-Created view that display hotel details -Corrected the form used in the search method -Corrected view that display the search results -Comments and general adjustments

+++ TesteBackendT4W/TesteBackendT4W/Models/SellingPricePerRoom.cs
@@ -1,8 +1,12 @@
 ﻿namespace TesteBackendT4W.Models
 {
+
+    /// <summary>
+    /// Room Attribute (obtained through API)  
+    /// </summary>
     public class SellingPricePerRoom
     {
         public string Currency { get; set; }
-        public int Value { get; set; }
+        public double Value { get; set; }
     }
 }
